backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
google-generativeai==0.7.2
anyio==4.4.0
python-dotenv==1.0.1
pandas==2.0.3
pathway>=0.10.0
